Make the proxy databsae agnostic at runtime
